state: split dashboardLayoutSlice into reducer/model modules; saves carry their fetch baseline
dashboardLayoutSlice.ts was 2,145 lines. This moves its reducers and helpers
into single-purpose modules next to it, keeps every exported action name and
the persisted layout shape, and shrinks the slice to the store wiring, the
fetch/save thunks and the extraReducers (568 lines):

  dashboardLayoutModel.ts        types, defaults, initial state
  dashboardLayoutGeometry.ts     grid/collision/placement helpers
  dashboardLayoutCardState.ts    one accessor registry for every CardType,
                                 z-order ledger rekey/remove/reconcile
  dashboardLayout{Agent,View,Browser,Workflow,Window,Canvas,Closed,Glow,
                  Reset}Reducers.ts

Behaviour changes, each with tests (node:test, 24 cases):

- Save authority. saveLayout now takes the per-dashboard baseline object the
  last successful fetch created (`saveAuthority`) and refuses a payload whose
  baseline is not the current one, or while a refetch for that dashboard is
  pending. A delayed or unmount save captured before a rejected/superseded
  fetch can no longer erase the server's layout. useLayoutSave and
  DashboardHeader's Share flush pass it from a selector.
- Fetch authority. Only the latest request generation may fulfill for the
  active dashboard; a late fetch cannot replace the active dashboard or be
  saved under its id, and a late rejection cannot revoke a newer baseline.
- Unknown persisted layout fields are read on fetch and written back on save
  instead of being dropped, so a newer build's fields survive an older one.
- z-order ledger is reconciled after a wholesale layout replace.

The geometry helpers read the viewport through viewportSize(), which falls
back exactly as the callers did (1440x900) when there is no window, so the
reducer tests run under node:test.
